WIP: large 146x146 mm arena with new layout
- Removed file format version checks from various components.
- Updated logging messages to include context in configuration loading.
- Adjusted CLI options to require arena directory input.
- Refactored GUI components to derive stage bounds from arena dimensions.
- Cleaned up unused includes and files related to file format versions.

+++ recorder/src/common/trackingControl.cpp
@@ -140,7 +140,7 @@ void motionControlRequestHandler(
             spdlog::critical(
                 "Motion stage request handler thread received unknown "
                 "request type: {}",
-                myRequest.requestType);
+                static_cast<int>(myRequest.requestType));
             throw std::runtime_error(
                 "Motion stage request handler thread received unknown "
                 "request type.");
